+++ src/edictor/arrayof.test.ts
@@ -1,6 +1,7 @@
 import { beforeEach, describe, expect, test } from '@jest/globals';
 import { ArrayOf, SetValueError, PushError } from './arrayof';
 import { AssertError, assert } from './util';
+import { Model, defineField } from '../edictor';
 
 describe('class ArrayOf', () => {
     let validators: any;
@@ -77,9 +78,17 @@ describe('class ArrayOf', () => {
     })
 
     test('ArrayOf().test()', () => {
+        class TestResult extends Model {};
+        TestResult.define({
+            test: defineField({required: true}).instance("string"),
+            valid: defineField().instance("object"),
+            invalid: defineField().instance("object")
+        })
+
         array = new ArrayOf('string', 'number');
         let result = array.test([1, true]);
-        console.log(result);
+
+        new TestResult(result);
     })
 
     test('ArrayOf().push()', () => {